fine tuning plant layer

+++ ch.obermuhlner.planetgen/src/ch/obermuhlner/planetgen/planet/Planet.java
@@ -148,7 +148,7 @@ public class Planet {
 
 	private Color convertTemperatureToColor(double temperature, double minTemperature, double maxTemperature) {
 		//double midTemperature = (maxTemperature - minTemperature) / 2 + minTemperature;
-		double midTemperature = Units.celsiusToKelvin((double) 0);
+		double midTemperature = Units.celsiusToKelvin(0);
 		return convertTemperatureToColor(temperature, minTemperature, midTemperature, maxTemperature);
 	}
 	
@@ -163,7 +163,7 @@ public class Planet {
 		}
 
 		// shows 0 degrees celsius as MAGENTA
-		double distanceToZero = MathUtil.smoothstep(0, 1, Math.abs(temperature - Units.celsiusToKelvin((double) 0)) / 0.5); 
+		double distanceToZero = MathUtil.smoothstep(0, 1, Math.abs(temperature - Units.celsiusToKelvin(0)) / 0.5); 
 		if (distanceToZero < 1.0) {
 			color = color.interpolate(Color.MAGENTA, 1.0 - distanceToZero);
 		}
